backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
dataclasses-json==0.6.7
exceptiongroup==1.3.1
finnhub-python==2.4.26
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
frozendict==2.4.7
frozenlist==1.8.0
gast==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.12.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langsmith==0.6.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
namex==0.1.0
nltk==3.9.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
peewee==3.19.0
platformdirs==4.5.1
praw==7.8.1
prawcore==2.4.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soupsieve==2.8.2
SQLAlchemy==2.0.45
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.3.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
vaderSentiment==3.3.2
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wrapt==2.0.1
yarl==1.22.0
yfinance==1.0
zstandard==0.25.0
